% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.data_subset_dataset.R
\name{add.new.series.imr}
\alias{add.new.series.imr}
\title{Add new IMR entries (dt_new_entries), now the same function as
`add.new.series.u5mr`}
\usage{
add.new.series.imr(dt_IMR, dt_new_entries, remove_old = FALSE)
}
\arguments{
\item{dt_IMR}{the IMR master dataset}

\item{dt_new_entries}{dt of new entries to be added}

\item{remove_old}{if TRUE, remove old entries, if FALSE, set old entries to
invisible and excluded}
}
\value{
the new dt_IMR
}
\description{
Still kept in case needed in the future in case need to differentiate IMR
from U5MR process
}
