% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.data_subset_dataset.R
\name{add.new.series.nmr}
\alias{add.new.series.nmr}
\title{Add new NMR entries (`dt_new_entries`)}
\usage{
add.new.series.nmr(dt_nmr, dt_new_entries)
}
\arguments{
\item{dt_nmr}{the NMR master dataset}

\item{dt_new_entries}{dt of new entries to be added}
}
\value{
the new dt_nmr_new
}
\description{
Add new NMR entries (`dt_new_entries`)
}
