% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.data_subset_dataset.R
\name{add.Original.Series.Name}
\alias{add.Original.Series.Name}
\title{Add an original series name}
\usage{
add.Original.Series.Name(dt_new_entries, ori_name)
}
\arguments{
\item{dt_new_entries}{dt of new entries to be added}

\item{ori_name}{value for Original.Series.Name}
}
\value{
dt_new_entries
}
\description{
Add `Original.Series.Name` column by the supplied `ori_name` argument
}
