% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.calculation.R
\name{calculate.4q1}
\alias{calculate.4q1}
\title{Calculate 4q1 from 5q0 and 1q0}
\usage{
calculate.4q1(q5, q1)
}
\arguments{
\item{q5}{5q0}

\item{q1}{1q0}
}
\value{
4q1
}
\description{
Calculate 4q1 from 5q0 and 1q0 or any part from the whole
e.g. NMR from IMR and PNMR
}
