% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.calculation.R
\name{calculate.arr}
\alias{calculate.arr}
\title{Get the logarithmic annual rate of reduction, expressed as a percentage
reduction}
\usage{
calculate.arr(dt, year1, year2)
}
\arguments{
\item{dt}{wide data by year, year1 and year2 shall be columns like 1990, 2000}

\item{year1}{year1 where year1 < year2}

\item{year2}{year2 where year1 < year2}
}
\value{
dt with added columns named like "1990-2000"
}
\description{
The formula for ARR is as follows: ARR = ln(rate2/rate1)/(t2-t1)*-100, where
t1 and t2 are years and t1<t2.
}
\examples{
\dontrun{
dt_final_ARR <- dt_final[Year\%in\%c(1990,2019)]
dt_final_ARR_w <- dcast(dt_final_ARR, ISO3Code ~ Year)
dt_final_ARR_w <- calculate.arr(dt_final_ARR_w, 1990, 2019)
}

}
