% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.data_engineering.R
\name{check.and.install.pkgs}
\alias{check.and.install.pkgs}
\title{check and install pkgs if missing}
\usage{
check.and.install.pkgs(pkgs)
}
\arguments{
\item{pkgs}{vector of packages}
}
\description{
check and install pkgs if missing
}
