% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.CC_funcs.R
\name{check.CC.profile.data}
\alias{check.CC.profile.data}
\title{Compare CC profile tables to "results.csv"}
\usage{
check.CC.profile.data(cc_dir, results_dir_list)
}
\arguments{
\item{cc_dir}{CC profile output directory: where the CC profiles are saved}

\item{results_dir_list}{list of all the results.csv files to read in and
compare}
}
\value{
NULL, print problems as messages, save comparison datasets
read CC xlsx profile data
}
\description{
Compare CC profile tables to all the "results.csv" files listed in
`results_dir_list` for all indicators enclosed.
}
