% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.CC_funcs.R
\name{compare.results.vs.cqt}
\alias{compare.results.vs.cqt}
\title{Compare the saved cqt vs results}
\usage{
compare.results.vs.cqt(dt_results, dt_cqt)
}
\arguments{
\item{dt_results}{obtained by `read.all.results.csv(results_dir_list)`}

\item{dt_cqt}{obtained by `get.dt.cqt(dir_cqt_files)`}
}
\description{
Should be the same as compare CC profiles vs results. It is used to check if
the cqt file and CC profile are both updated in the same time
}
