% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{country_info}
\alias{country_info}
\title{country_info: country.info.CME.csv}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 196 rows and 381 columns.}
\usage{
country_info
}
\description{
country_info: country.info.CME.csv
}
\keyword{dataset}
