% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{default_label}
\alias{default_label}
\title{A place holder, if supplied in global environment it will be used in the
\code{\link{get.match}} function}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
default_label
}
\description{
A place holder, if supplied in global environment it will be used in the
\code{\link{get.match}} function
}
\keyword{labels}
