% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{default_label_1}
\alias{default_label_1}
\title{some pre-stored labels, will update from time to time}
\format{An object of class \code{list} of length 21.}
\usage{
default_label_1
}
\description{
some pre-stored labels, will update from time to time
}
\keyword{labels}
