% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{dir_list}
\alias{dir_list}
\title{dir_list: useful dropbox dir list}
\format{
An object of class \code{list} of length 3.
}
\usage{
dir_list
}
\description{
dir_list: useful dropbox dir list
}
\keyword{directory}
