% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_sync_off.R
\name{dropbox_project_sync_off}
\alias{dropbox_project_sync_off}
\title{Let Dropbox untrack `Rproj user` folder, solve Dropbox conflicts with
`Rproj.user`}
\usage{
dropbox_project_sync_off()
}
\description{
Revised base on:
https://community.rstudio.com/t/dropbox-conflicts-with-rproj-user/54059
}
