% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{final_dir_list}
\alias{final_dir_list}
\title{final_dir_list: saved dir list of final results, could also be obtained using
`load.final_dir()`}
\format{
An object of class \code{list} of length 6.
}
\usage{
final_dir_list
}
\description{
final_dir_list: saved dir list of final results, could also be obtained using
`load.final_dir()`
}
\keyword{directory}
