% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.CC_funcs.R
\name{get.0_4.sources}
\alias{get.0_4.sources}
\title{Get survey names for Table Under 5}
\usage{
get.0_4.sources(dt_U5MR, dt_IMR, dt_NMR)
}
\arguments{
\item{dt_U5MR}{U5MR master dataset}

\item{dt_IMR}{IMR}

\item{dt_NMR}{NMR}
}
\value{
list of 2 datasets: surveys sources and VR sources
}
\description{
Get survey names for Table Under 5
}
