% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.CC_funcs.R
\name{get.5_24.sources}
\alias{get.5_24.sources}
\title{Get survey names (sources) for 5 to 24}
\usage{
get.5_24.sources(dt5_14, dt15_24)
}
\arguments{
\item{dt5_14}{dt 5-14}

\item{dt15_24}{dt 15-24}
}
\value{
list of 3 datasets: sources for 5-14, for 15-24 and vr
}
\description{
Get survey names (sources) for 5 to 24
}
