% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helper.R
\name{get.CME.data}
\alias{get.CME.data}
\title{get three types (Under-five Infant Neonatal) from Rates & Deaths summary}
\usage{
get.CME.data(year_range = c(2000:2018), get_what = NULL)
}
\arguments{
\item{year_range}{a vector of years we want, default to 2000:2018}

\item{get_what}{"Deaths" or "Rate", default to "Rate": get the three CME rate}
}
\value{
dt of ISO3, UNcode, year, Under-five, Infant, Neonatal, one row for each country each year in year_range
}
\description{
get three types (Under-five Infant Neonatal) from Rates & Deaths summary
}
\examples{
get.CME.data(year_range = c(2016:2018))
}
