% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_engineering.R
\name{get.CME.estimates.long}
\alias{get.CME.estimates.long}
\title{Read in IGME 2019 estimates, using CME_2019_all_data
01/24/2020}
\usage{
get.CME.estimates.long(
  ind = c("U5MR", "NMR"),
  get = "both",
  c_iso = NULL,
  c_name = "Mozambique",
  year_range = 1990:2018
)
}
\arguments{
\item{ind}{choose indicators among U5MR, NMR, and IMR}

\item{get}{both, rate, or death}

\item{c_iso}{get country by ISO, will overwrite c_name if provided}

\item{c_name}{country by official name}

\item{year_range}{a vector, e.g. 1990:2018}
}
\value{
a data.table
}
\description{
Read in IGME 2019 estimates, using CME_2019_all_data
01/24/2020
}
\examples{
dt_cme_long <- get.CME.estimates.long(ind = "U5MR", c_name = "Mozambique", year_range = 2016:2018)
}
