% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{get.CME.ind.data}
\alias{get.CME.ind.data}
\title{Load country indicator from CME output data}
\usage{
get.CME.ind.data(
  shortind = "U5MR",
  round = NULL,
  median_only = TRUE,
  total_sex = TRUE
)
}
\arguments{
\item{shortind}{Name of `Shortind`}

\item{round}{default to NULL, IGME round on Dropbox, must be supplied}

\item{median_only}{default to TRUE, return medians}

\item{total_sex}{default to TRUE, return sex = Total}
}
\description{
Load country indicator from CME output data
}
