% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_engineering.R
\name{get.CME.UI.data}
\alias{get.CME.UI.data}
\title{return different format of data using the final aggregate results country
summary. This is a simpler early version I wrote}
\usage{
get.CME.UI.data(
  c_iso = NULL,
  year_range = c(1990:2018),
  ind = "all",
  get = "rate",
  idvars = c("OfficialName", "ISO3Code"),
  dir_file = NULL,
  match_ind = NULL,
  format = "long",
  round_digit = 1L,
  use_sample_data = FALSE
)
}
\arguments{
\item{c_iso}{country iso, if NULL, returns all isos, default to NULL}

\item{year_range}{a vector of years, default to 1990: 2018}

\item{ind}{be either "U5MR", "NMR", "IMR", "all", default to "all"}

\item{get}{get either "rate", "death", or "both", default to "rate"}

\item{idvars}{default to "OfficialName, ISO3Code", what id vars you want to
include}

\item{dir_file}{allow a different dataset to be read}

\item{match_ind}{a different rules to create variable to read using
`get.match`}

\item{format}{Choose format among raw, long, wide_year, wide_ind, and
wide_get, default to "long". All the wide-format just decasts the
long-format data}

\item{round_digit}{round estimates, default to 1}

\item{use_sample_data}{use a build-in sample data for fast evaluating the
function}
}
\value{
a data.table
}
\description{
return different format of data using the final aggregate results country
summary. This is a simpler early version I wrote
}
\examples{
dt_t <- get.CME.UI.data(c_iso = "AFG", year_range = 2000:2018,
ind = "all", use_sample_data = TRUE)
dt_f <- get.CME.UI.data(c_iso = "AFG", year_range = 1960:2018,
ind = c("U5MR", "IMR"), format = "wide_year", use_sample_data = TRUE)


}
