% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{get.country.info.CME}
\alias{get.country.info.CME}
\title{Load the "country.info.CME"}
\usage{
get.country.info.CME(year0, work_dir = NULL)
}
\arguments{
\item{year0}{IGME round on Dropbox, until 2023}

\item{work_dir}{for IGME round on Dropbox, provide work_dir}
}
\value{
dataset of country info
}
\description{
Creates UNICEFReportRegion from UNICEFReportRegion1 and UNICEFReportRegion2
}
