% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.data_engineering.R
\name{get.country.info}
\alias{get.country.info}
\title{Load the country info file, requires dt object `country_info`}
\usage{
get.country.info()
}
\value{
dataset of country info
}
\description{
Creats UNICEFReportRegion from UNICEFReportRegion1 and UNICEFReportRegion2
}
