% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.CC_funcs.R
\name{get.cqt.from.results}
\alias{get.cqt.from.results}
\title{Make cqt file from `results.csv` file}
\usage{
get.cqt.from.results(
  output_dir,
  mcmc.meta_filename = "mcmc.meta.rda",
  results_filename = "results.csv",
  res.cqt_filename = "res.cqt.Lw.rda"
)
}
\arguments{
\item{output_dir}{where to find the results.csv}

\item{mcmc.meta_filename}{default to "mcmc.meta.rda"}

\item{results_filename}{default to "results.csv", the file to read}

\item{res.cqt_filename}{default to "res.cqt.Lw.rda", file name for saving}
}
\description{
Read `results.csv` and `mcmc.meta` files in the output dir and make res.cqt.Lw.rda
}
