% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.data_subset_dataset.R
\name{get.dir_IMR}
\alias{get.dir_IMR}
\title{Get the IMR master dataset directory}
\usage{
get.dir_IMR(dir_IGME = get.IGMEinput.dir(2020))
}
\arguments{
\item{dir_IGME}{The directory to IGME input folder, e.g. ".../2020 Round
Estimation/Code/input/"}
}
\value{
file path to the master dataset
}
\description{
Get the IMR master dataset directory
}
