% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{get.dir_latest_file}
\alias{get.dir_latest_file}
\title{Get the file directory with latest date in the filename}
\usage{
get.dir_latest_file(dir_folder, pattern_to_match)
}
\arguments{
\item{dir_folder}{The directory to search for files}

\item{pattern_to_match}{Pattern used to match filename}
}
\value{
file path to the dataset
}
\description{
Get the file directory with latest date in the filename
}
