% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.data_subset_dataset.R
\name{get.dir_NMR}
\alias{get.dir_NMR}
\title{Get the NMR master dataset directory}
\usage{
get.dir_NMR(y5 = FALSE)
}
\arguments{
\item{y5}{to get the 5-year dataset or not}
}
\value{
file path to the master dataset
}
\description{
Compare to \code{\link{get.dir_U5MR}}, there is need to supply dir_IGME since
the dataset location is fixed at "/NMR/data"
}
