% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_find_directories.R
\name{get.dir_U5MR}
\alias{get.dir_U5MR}
\title{Get the U5MR master dataset directory}
\usage{
get.dir_U5MR(dir_IGME = get.IGMEinput.dir(2021))
}
\arguments{
\item{dir_IGME}{The directory to IGME input folder, e.g. ".../2020 Round
Estimation/Code/input/", could be obtained using
\code{\link{get.IGMEinput.dir}}}
}
\value{
file path to the master dataset
}
\description{
Get the U5MR master dataset directory
}
