% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.CC_funcs.R
\name{get.dt.cqt}
\alias{get.dt.cqt}
\title{Collect data from the cqt file for plotting}
\usage{
get.dt.cqt(cqt_dir)
}
\arguments{
\item{cqt_dir}{where all the `res.cqt.rds` files are saved}
}
\value{
dt_cqt
}
\description{
`get.dt.cqt` loads all cqt files in the folder as one long-format dataset
}
