% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_engineering.R
\name{get.dt.wide}
\alias{get.dt.wide}
\title{Get wide format for dataset}
\usage{
get.dt.wide(dt_long)
}
\arguments{
\item{dt_long}{a long-format data with "Country", "Year", "Indicator","Estimate", "Lower_Bound", "Upper_Bound"}
}
\value{
a wide-format data
}
\description{
Get wide format for dataset
}
\examples{
dt_cme_wide <- get.dt.wide(get.CME.estimates.long(c_name = "China", year_range = 2018))
}
