% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_find_directories.R
\name{get.file.name}
\alias{get.file.name}
\title{Show all file directories within the file directory `dir_file` and matched by
pattern `pattern0`}
\usage{
get.file.name(dir_file, pattern0, full_name = TRUE)
}
\arguments{
\item{dir_file}{directory}

\item{pattern0}{string to match file names}

\item{full_name}{list.files(full.names), if TRUE (default) returns full
directories, if FALSE, return only the file names}
}
\value{
vector of matched file directories
}
\description{
Search only the files in the folder, match by `pattern0`, the search is not
recursive.
}
