% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_find_directories.R
\name{get.IGMEfig.dir}
\alias{get.IGMEfig.dir}
\title{Get "fig" dir for a given year}
\usage{
get.IGMEfig.dir(year)
}
\arguments{
\item{year}{YYYY}
}
\value{
directory to fig folder
}
\description{
If `year` is 2020, returns the directory to fig folder in the 2020 Round
Estimation Dropbox folder
}
