% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{get.label}
\alias{get.label}
\title{a label function to relabel certain variable names}
\usage{
get.label(x, new_list = NULL, no_line_break = FALSE)
}
\arguments{
\item{x}{a element or a vector}

\item{new_list}{if you supply a new list the function will use instead of the default_labels}

\item{no_line_break}{to remove \n from the string}
}
\value{
updated labels as character vector
}
\description{
a label function to relabel certain variable names
}
