% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{get.live.birth}
\alias{get.live.birth}
\title{Load the "data_livebirths.csv", add ISO3Code and country names}
\usage{
get.live.birth(year0, work_dir)
}
\arguments{
\item{year0}{IGME round on Dropbox, until 2023}

\item{work_dir}{for IGME round on Dropbox, provide work_dir}
}
\value{
dataset of live birth
}
\description{
Load the "data_livebirths.csv", add ISO3Code and country names
}
