% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.data_engineering.R
\name{get.match}
\alias{get.match}
\title{A label function to replace values by a given list in a variable}
\usage{
get.match(x, new_list = NULL, no_line_break = FALSE, show_no_match = FALSE)
}
\arguments{
\item{x}{a element or a vector}

\item{new_list}{if you supply a new list the function will use instead of the
default_labels}

\item{no_line_break}{default to FALSE, if TRUE will remove line break from
the string}

\item{show_no_match}{default to FALSE, if TRUE will message unmatched
elements}
}
\value{
an updated vector as character
}
\description{
You can provide a __new_list__ to define the values you wish to change in
this variable. Values not revised in the given list will be kept
}
