% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.CC_funcs.R
\name{get.new.sourceID}
\alias{get.new.sourceID}
\title{Format the legend_ID for the master datasets}
\usage{
get.new.sourceID(data = dt_cme)
}
\arguments{
\item{data}{datasets}
}
\value{
subsetted dataset with revised Series.Name and new column `legend_ID`
}
\description{
Used in \code{\link{get.0_4.sources}} and \code{\link{get.5_24.sources}} to
create sources for template input
}
