% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{get.numeric.date}
\alias{get.numeric.date}
\title{Transform date into numeric numbers like 2020.55}
\usage{
get.numeric.date(date0)
}
\arguments{
\item{date0}{date for example: 2020-01-01}
}
\value{
numeric date for example: 2020.014
}
\description{
Transform date into numeric numbers like 2020.55
}
