% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{get.pop.from.CME}
\alias{get.pop.from.CME}
\title{Extract long-format pop0 and pop1-4 from country.info.CME}
\usage{
get.pop.from.CME(dc)
}
\arguments{
\item{dc}{country.info.CME}
}
\value{
long-format ISO3Code, year, pop0 and pop1-4
}
\description{
Extract long-format pop0 and pop1-4 from country.info.CME
}
