% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_find_directories.R
\name{get.raw.dir}
\alias{get.raw.dir}
\title{Get raw file directory from `Output CMRJack` folder}
\usage{
get.raw.dir(cname, surveytype = "DHS", year = NULL)
}
\arguments{
\item{cname}{country name}

\item{surveytype}{folder names like "DHS", "MICS", "NDHS",...}

\item{year}{year of the survey, e.g. 2015}
}
\value{
xlsx file directory
}
\description{
Get raw file directory from `Output CMRJack` folder
}
\examples{
\dontrun{
get.raw.dir("Zimbabwe", "DHS", 2015)
}
}
