% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.CC_funcs.R
\name{get.table.by.iso}
\alias{get.table.by.iso}
\title{Get CC profile data by country}
\usage{
get.table.by.iso(iso0, ind_wanted = c("10q15", "10q5", "IMR", "NMR", "U5MR"))
}
\arguments{
\item{iso0}{ISO3Code}

\item{ind_wanted}{indicators used in the profile, default to `c("10q15",
"10q5", "IMR", "NMR", "U5MR")`}
}
\value{
formatted table by country
}
\description{
Get the table that fits into the CC profile excel template. Fit the format of
the template with blank rows between indicators. Follow the order of median,
lower and upper. Only the values (not row/col names) are put into the
template later
}
