% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{get.workdir.dropbox}
\alias{get.workdir.dropbox}
\title{Get IGME "Code" dir for a given year from Dropbox}
\usage{
get.workdir.dropbox(year = 2023)
}
\arguments{
\item{year}{YYYY}
}
\value{
directory to folder
}
\description{
Get IGME "Code" dir for a given year from Dropbox
}
