% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{get.workdir.sharepoint}
\alias{get.workdir.sharepoint}
\title{GReturn the IGME "Code" dir on SharePoint}
\usage{
get.workdir.sharepoint(year = 2024)
}
\arguments{
\item{year}{YYYY}
}
\value{
directory to "Code" folder
}
\description{
If `year` is 2020, returns the directory to Code folder in the 2020 Round
Estimation Dropbox folder
}
