% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.directories.R
\name{get_os}
\alias{get_os}
\title{check system, return . Used as nternal function.}
\usage{
get_os()
}
\value{
"Windows", "OSX", or "Linus"
}
\description{
check system, return . Used as nternal function.
}
