% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{hiv.iso}
\alias{hiv.iso}
\title{hiv.iso: the 17 hiv isos}
\format{
An object of class \code{character} of length 17.
}
\usage{
hiv.iso
}
\description{
hiv.iso: the 17 hiv isos
}
\keyword{iso}
