% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.data_engineering.R
\name{load.final_dir}
\alias{load.final_dir}
\title{Return a saved dir_list for total and sex-specific Rates & Deaths_Country
Summary.csv}
\usage{
load.final_dir()
}
\value{
list of results file directories on Dropbox
}
\description{
`load.final_dir` returns file paths of Rates & Deaths_Country Summary.csv
from Aggregate results (final) folders
}
