% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.directories.R
\name{load.IGMEinput.dir}
\alias{load.IGMEinput.dir}
\title{Load the IGME "input" directories}
\usage{
load.IGMEinput.dir()
}
\value{
dir_IGMEinput_list: a list of directories of UN IGME Data/YYYY Round
  Estimations/Code/input
}
\description{
Load the IGME "input" directories
}
