% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{Rates_Deaths_Country_Summary_2020_UI}
\alias{Rates_Deaths_Country_Summary_2020_UI}
\title{Rates_Deaths_Country_Summary_2020_UI: The 2020 final estimates with UI
Please refer to childmortality.org for details}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 588 rows and 427 columns.
}
\usage{
Rates_Deaths_Country_Summary_2020_UI
}
\description{
Rates_Deaths_Country_Summary_2020_UI: The 2020 final estimates with UI
Please refer to childmortality.org for details
}
\keyword{dataset}
