% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.data_subset_dataset.R
\name{rbinddataset}
\alias{rbinddataset}
\title{Row-bind two datasets, check duplicated keys and set order}
\usage{
rbinddataset(dt_master, dt_new)
}
\arguments{
\item{dt_master}{master dataset}

\item{dt_new}{new entries}
}
\value{
`dt1` as `rbind(dt_master, dt_new)`
}
\description{
Row-bind two datasets, check duplicated keys and set order
}
