% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_add_new_data_series.R
\name{rbinddatasetNMR}
\alias{rbinddatasetNMR}
\title{Row-bind two datasets for NMR with slight changes, check duplicated keys, set
order}
\usage{
rbinddatasetNMR(dt_master, dt_new)
}
\arguments{
\item{dt_master}{master dataset}

\item{dt_new}{new entries}
}
\value{
`dt1` as `rbind(dt_master, dt_new)`
}
\description{
Row-bind two datasets for NMR with slight changes, check duplicated keys, set
order
}
