% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.CC_funcs.R
\name{read.all.results.csv}
\alias{read.all.results.csv}
\title{Read all "results.csv"}
\usage{
read.all.results.csv(results_dir_list, year_range0 = 1931:2019)
}
\arguments{
\item{results_dir_list}{results_dir_list list of all the results.csv files to
read in and compare}

\item{year_range0}{year range, default to all years 1931 to 2019}
}
\value{
dt_results_2020, saved as csv file in folder Results_Data too
}
\description{
Using \code{\link{read.results.csv.file}} to read all "results.csv" files
}
