% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.data_engineering.R
\name{read.country.summary}
\alias{read.country.summary}
\title{function to read "Rates & Deaths_Country Summary.csv" and output long format}
\usage{
read.country.summary(dir_dt_cs, year_wanted = NULL, sex = NULL)
}
\arguments{
\item{dir_dt_cs}{directory to Rates & Deaths_Country Summary.csv}

\item{year_wanted}{default to null, supply e.g. 1990:2019}

\item{sex}{default to NULL, will determine from file dir}
}
\description{
function to read "Rates & Deaths_Country Summary.csv" and output long format
}
