% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.data_engineering.R
\name{read.region.summary}
\alias{read.region.summary}
\title{function to read regional summary and output long format}
\usage{
read.region.summary(
  dir_dt_cs,
  year_wanted = NULL,
  sex = NULL,
  add_regional_grouping = FALSE
)
}
\arguments{
\item{dir_dt_cs}{directory to e.g. paste0("Rates & Deaths_UNICEFRegion.csv")}

\item{year_wanted}{default to null, supply e.g. 1990:2019}

\item{sex}{default to NULL, will determine from file dir}

\item{add_regional_grouping}{if add `Regional_Grouping` column}
}
\description{
function to read regional summary and output long format
}
