% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{read.region.summary.SBproj}
\alias{read.region.summary.SBproj}
\title{sometimes maybe useful, to read the wide-year format SBR output}
\usage{
read.region.summary.SBproj(dir_SBR_wide_year, value_name0 = "Median")
}
\arguments{
\item{dir_SBR_wide_year}{directory to the wide-year format SBR output}

\item{value_name0}{default to "Median"}
}
\description{
sometimes maybe useful, to read the wide-year format SBR output
}
