% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.CC_funcs.R
\name{read.results.csv.file}
\alias{read.results.csv.file}
\title{Read one results.csv file and reformat into long-format}
\usage{
read.results.csv.file(
  dt_dir,
  year_range = 1931:2019,
  q = c("Lower", "Median", "Upper"),
  sex = NULL
)
}
\arguments{
\item{dt_dir}{the single directory to a results.csv file}

\item{year_range}{year range desired, default to `1931:2019`}

\item{q}{quantile desired, default to `c("Lower", "Median", "Upper")`}

\item{sex}{default to NULL, Sex is determined from `dt_dir`, unless specified}
}
\value{
long-format dt containing all indicators
}
\description{
Read one results.csv file and reformat into long-format
}
