% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{read.results.csv}
\alias{read.results.csv}
\title{Read one results.csv file and reformat into long-format}
\usage{
read.results.csv(
  dt_dir,
  year_range = 1931:2030,
  q = c("Lower", "Median", "Upper"),
  sex = NULL
)
}
\arguments{
\item{dt_dir}{the single directory to a results.csv file}

\item{year_range}{year range desired, default to all years: `1931:2030`}

\item{q}{quantile desired, default to `c("Lower", "Median", "Upper")`}

\item{sex}{default to NULL, sex is determined from `dt_dir`, unless specified}
}
\value{
long-format results.csv, all values in the "value" column
}
\description{
Read one results.csv file and reformat into long-format
}
