% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{read.SBproj}
\alias{read.SBproj}
\title{Read the wide-year format SBR output (country or region)}
\usage{
read.SBproj(dir_SBR_wide_year, id_wanted = NULL, value_name0 = "value")
}
\arguments{
\item{dir_SBR_wide_year}{directory to the wide-year format SBR output}

\item{id_wanted}{should be either "ISO3Code" or "Region", default to NULL: so
must be provided}

\item{value_name0}{default to "value", what name to mark the value}
}
\description{
Read the wide-year format SBR output (country or region)
}
