% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.data_subset_dataset.R
\name{revise.age.group}
\alias{revise.age.group}
\title{Remove 25-34 age group in `dt_new_entries`}
\usage{
revise.age.group(dt_new_entries)
}
\arguments{
\item{dt_new_entries}{dt of new entries to be added}
}
\description{
Internal function used by `add.new.series` functions
}
