% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{revise.path}
\alias{revise.path}
\title{Adjust the file dir if the lash is not right or the Dropbox username is not
right}
\usage{
revise.path(dir0)
}
\arguments{
\item{dir0}{file directory not output for now}
}
\description{
Adjust the file dir if the lash is not right or the Dropbox username is not
right
}
