% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.CC_funcs.R
\name{save.CME.CC.profile}
\alias{save.CME.CC.profile}
\title{Save CMR CC profile by country}
\usage{
save.CME.CC.profile(
  iso0,
  template = "IGME CC Template.xlsx",
  dir_save = "AfterCC"
)
}
\arguments{
\item{iso0}{ISO3Code}

\item{template}{directory to template, default to project root folder}

\item{dir_save}{directory to save output, default to folder "AfterCC",
folders are created if doesn't exist}
}
\value{
NULL, save xlsx in `dir_save`
}
\description{
`save.CME.CC.profile` writes excel template using the `XLSX` package
}
