% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_engineering.R
\name{search.for.file}
\alias{search.for.file}
\title{search all the folders in `target.dir`, and list files containing the `file_name_string`}
\usage{
search.for.file(target.dir, file_name_string, full_path = FALSE)
}
\arguments{
\item{target.dir}{target directory}

\item{file_name_string}{e.g. "data_U5MR"}

\item{full_path}{full path or not}
}
\description{
search all the folders in `target.dir`, and list files containing the `file_name_string`
}
