% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{shortind_indicator}
\alias{shortind_indicator}
\title{List of shortind_indicator: match shortind to full indicator names}
\format{
An object of class \code{list} of length 28.
}
\usage{
shortind_indicator
}
\description{
List of shortind_indicator: match shortind to full indicator names
}
\keyword{labels}
