% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{UNICEF_colors}
\alias{UNICEF_colors}
\title{UNICEF_colors contains the UNICEF color palette including UNICEF blue,
Sequential, Diverging, Qualitative, and grays scheme}
\format{
An object of class \code{list} of length 5.
}
\usage{
UNICEF_colors
}
\description{
UNICEF_colors contains the UNICEF color palette including UNICEF blue,
Sequential, Diverging, Qualitative, and grays scheme
}
\details{
UNICEF_colors <- list(
    "UNICEF_blue" = "#1CABE2",
    # Sequential scheme
    "cp_UNICEF_seq" = c( "#002759", "#00377D", "#0058AB","#0083CF","#1CABE2","#69DBFF",
                         "#A3EAFF","#CFF4FF"),
    # Diverging scheme
    "cp_UNICEF_div" = c("#002759", "#0058AB", "#1CABE2", "#69DBFF", "#CFF4FF", "#FFF09C",
                        "#FFC20E", "#F26A21", "#E2231A", "#B50800"),
    # Qualitative schemes
    "cp_UNICEF_qua" = c("#0058AB", "#1CABE2", "#00833D", "#80BD41", "#6A1E74", "#961A49",
                        "#E2231A", "#F26A21", "#FFC20E", "#FFF09C"),
    # grays
    "cp_UNICEF_grays" = c("#333333", "#6B6B6B","#B3B3B3","#D7D7D7","#E8E8E8")
    )
}
\keyword{color}
\keyword{scheme}
