% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.data_engineering.R
\name{upper.first.letter}
\alias{upper.first.letter}
\title{Capitalize first letter of each word in the vector}
\usage{
upper.first.letter(y)
}
\arguments{
\item{y}{vector of strings}
}
\value{
a vector of strings with first letter capitalized
}
\description{
Capitalize first letter of each word in the vector
}
\examples{
upper.first.letter(c("aa","bb","cc"))
}
