% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helper.R
\name{xlsx.writeMultipleData}
\alias{xlsx.writeMultipleData}
\title{function to write several dataset into excel file (one dataset each sheet) with table format}
\usage{
xlsx.writeMultipleData(file, ..., sheet_name = NULL)
}
\arguments{
\item{file}{the path to the output file}

\item{...}{a list of data to write to the workbook}

\item{sheet_name}{must be a vector of sheet names you want, if length != length of object, it will be ignored}
}
\description{
revised based on http://www.sthda.com/english/wiki/r-xlsx-package-a-quick-start-guide-to-manipulate-excel-files-in-r
}
\examples{
xlsx.writeMultipleData("myworkbook.xlsx", mtcars, Titanic, sheet_name = c("dt1", "dt2"))

}
