% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.calculation.R
\name{calculate.5q0}
\alias{calculate.5q0}
\title{Calculate 5q0 from 1q0 and 4q1}
\usage{
calculate.5q0(q1, q4)
}
\arguments{
\item{q1}{1q0}

\item{q4}{4q1}
}
\value{
5q0
}
\description{
Calculate 5q0 from 1q0 and 4q1
}
