% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{check.and.install.pkgs}
\alias{check.and.install.pkgs}
\title{Check and install packages if missing}
\usage{
check.and.install.pkgs(pkgs)
}
\arguments{
\item{pkgs}{vector of packages}
}
\description{
Check and install packages if missing
}
