% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{create.IGME.key}
\alias{create.IGME.key}
\title{Create IGME_Key column}
\usage{
create.IGME.key(
  dt0,
  add_Indicator = FALSE,
  add_VR_series_name = FALSE,
  add_series_type = TRUE
)
}
\arguments{
\item{dt0}{dataset}

\item{add_Indicator}{default to `FALSE`, if `TRUE` will add indicator in the
end of `IGME_Key`, keep another `IGME_Key_s` column which doesn't have
indicator in the end}

\item{add_VR_series_name}{do we want to include (S)VR Series.Name and
Series.Year in the key}

\item{add_series_type}{do want to add Series.Type in the end of the key}
}
\value{
dt0 dataset with added column `IGME_Key`
}
\description{
Create column `IGME_Key` in the format of
ISO3Code-Series.Year-Series.Name-Series.Category,
e.g."TZA-1988-Census-Indirect", "RWA-2017-Malaria Indicator Survey-Indirect".
Extra strings like "Preliminary" or "MM/NN adjusted" are removed
}
