% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{default_label}
\alias{default_label}
\title{a place holder, supply and it will be used in the `get.match` function}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
default_label
}
\description{
a place holder, supply and it will be used in the `get.match` function
}
\keyword{labels}
