% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.directories.R
\name{get.dir_gender}
\alias{get.dir_gender}
\title{Get the sex-specific master dataset directory}
\usage{
get.dir_gender(dir_IGME_gender = NULL, plotting = TRUE)
}
\arguments{
\item{dir_IGME_gender}{default to "/Dropbox/CMEgender2015/Database"}

\item{plotting}{to get the dataset for plotting (if TRUE) or dataset for
modeling (if FALSE)}
}
\value{
file path to the master dataset
}
\description{
Compare to \code{\link{get.dir_U5MR}}, there is need to supply dir_IGME since
the dataset location is fixed at "/CMEgender2015/Database"
}
