% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{get.dir_IMR}
\alias{get.dir_IMR}
\title{Get the IMR master dataset directory}
\usage{
get.dir_IMR(workdir)
}
\arguments{
\item{workdir}{The directory to IGME input folder, e.g. ".../2020 Round
Estimation/Code/input/"}
}
\value{
file path to the master dataset
}
\description{
Get the IMR master dataset directory
}
