% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{get.dir_U5MR}
\alias{get.dir_U5MR}
\title{Get the U5MR master dataset directory}
\usage{
get.dir_U5MR(workdir, pattern_to_match = "data_U5MR")
}
\arguments{
\item{workdir}{The directory to IGME Code folder, e.g. ".../202x Round
Estimation/Code"}

\item{pattern_to_match}{default to "data_U5MR", but can be used generally}
}
\value{
file path to the master dataset
}
\description{
Get the U5MR master dataset directory
}
