% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{get.label}
\alias{get.label}
\title{a label function to relabel certain variable names}
\usage{
get.label(x, new_list = NULL, no_line_break = FALSE)
}
\arguments{
\item{x}{a element or a vector}

\item{new_list}{if you supply a new list the function will use instead of the
default_labels}

\item{no_line_break}{to remove linebreak from the string}
}
\value{
updated labels as character vector
}
\description{
You can provide a __new_list__ to define the labels. If any label is not
provided, the function will just return the original value
}
