% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{get.live.birth}
\alias{get.live.birth}
\title{Load the "data_livebirths.csv", add ISO3Code and country names}
\usage{
get.live.birth(year0 = 2021)
}
\arguments{
\item{year0}{IGME round, e.g. 2021}
}
\value{
dataset of live birth
}
\description{
Load the "data_livebirths.csv", add ISO3Code and country names
}
