% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.directories.R
\name{get.numeric.date}
\alias{get.numeric.date}
\title{return numeric date}
\usage{
get.numeric.date(date0)
}
\arguments{
\item{date0}{date for example: 2020-01-01}
}
\value{
numeric date for example: 2020.014
}
\description{
return numeric date
}
