% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{ggsave.figs}
\alias{ggsave.figs}
\title{use ggsave to save figs in both png and pdf, default to 600dpi, in folder "figure/"}
\usage{
ggsave.figs(file_name, myplot, width = 8, height = 6,
  folder_name = "figure")
}
\arguments{
\item{file_name}{the name of the plot file, e.g, "plot1"}

\item{myplot}{ggplot2 object}

\item{width}{width for ggsave}

\item{height}{height for ggsave}

\item{folder_name}{the relative dir for here()}
}
\description{
use ggsave to save figs in both png and pdf, default to 600dpi, in folder "figure/"
}
