% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.directories.R
\name{leap_year}
\alias{leap_year}
\title{Internal function: Check if `date` is a leap year}
\usage{
leap_year(date)
}
\arguments{
\item{date}{date}
}
\description{
Internal function: Check if `date` is a leap year
}
