% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{read.all.results.csv}
\alias{read.all.results.csv}
\title{Read and bind all "results.csv" into one file using `rbindlist`}
\usage{
read.all.results.csv(
  results_dir_list,
  year_range0 = 1931:2030,
  value_name = "Results"
)
}
\arguments{
\item{results_dir_list}{results_dir_list list of all the results.csv files to
read in and compare}

\item{year_range0}{year range, default to years `1931:2021`}

\item{value_name}{name of the value, default to "Results"}
}
\value{
long-format results.csv for all indicators, all values in the
  `value_name` column
}
\description{
Use \code{\link{read.results.csv}} to read all "results.csv" files
}
