% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{read.country.summary}
\alias{read.country.summary}
\title{Read "Rates & Deaths_Country Summary.csv" and output long format}
\usage{
read.country.summary(dir_file, year_range = NULL, sex = NULL)
}
\arguments{
\item{dir_file}{directory to Rates & Deaths_Country Summary.csv}

\item{year_range}{default to null, supply e.g. 1990:2019}

\item{sex}{default to NULL, will determine from file dir}
}
\description{
This function is a simpler version of \code{\link{get.CME.UI.data}} and is
more straightforward to use: without many arguments, only output long format.
The arguments are aligned so this function could be replaced by
\code{\link{get.CME.UI.data}} in the code if needed
}
