% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{read.region.summary}
\alias{read.region.summary}
\title{Read regional summary and output long format}
\usage{
read.region.summary(
  dir_file,
  year_range = NULL,
  sex = NULL,
  add_regional_grouping = FALSE
)
}
\arguments{
\item{dir_file}{directory to e.g. paste0("Rates & Deaths_UNICEFRegion.csv")}

\item{year_range}{default to null, supply e.g. 1990:2019}

\item{sex}{default to NULL, will determine from file dir}

\item{add_regional_grouping}{if add `Regional_Grouping` column}
}
\description{
Read regional summary and output long format
}
