% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_add_new_data_series.R
\name{revise.age.group}
\alias{revise.age.group}
\title{Remove 25-34 age group in `dt_new_entries`}
\usage{
revise.age.group(dt_new_entries)
}
\arguments{
\item{dt_new_entries}{dt of new entries to be added}
}
\value{
dt_new_entries with 25-34 and 10-19 removed
}
\description{
function used by `add.new.series` functions, remove unwanted age groups
}
