% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.data_engineering.R
\name{roundoff}
\alias{roundoff}
\title{A rounding function that round off numbers in the conventional way}
\usage{
roundoff(x, digits = 2)
}
\arguments{
\item{x}{the number}

\item{digits}{digits, default to 2}
}
\value{
rounded numeric vector
}
\description{
Instead of in R where round(0.5) = 0, roundoff(0.5, 0) = 1
}
