% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_engineering.R
\name{roundoff}
\alias{roundoff}
\title{a round function that Round off numbers in the conventional way instead of the R round
In R, round(0.5) = 0}
\usage{
roundoff(x, digits = 2)
}
\arguments{
\item{x}{the number}

\item{digits}{digits, default to 2}
}
\description{
a round function that Round off numbers in the conventional way instead of the R round
In R, round(0.5) = 0
}
