% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.data_engineering.R, R/3.data_subset_dataset.R
\name{roundoff}
\alias{roundoff}
\title{a rounding function that Round off numbers in the conventional way instead of
the R round In R, round(0.5) = 0}
\usage{
roundoff(x, digits = 2)

roundoff(x, digits = 2)
}
\arguments{
\item{x}{the number}

\item{digits}{digits, default to 2}
}
\value{
a rounded vector
}
\description{
a rounding function that Round off numbers in the conventional way instead of
the R round In R, round(0.5) = 0

a round function that Round off numbers in the conventional way instead of the R round
In R, round(0.5) = 0
}
