% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_read_data.R
\name{search.for.file}
\alias{search.for.file}
\title{Search for file paths matched by part of the file name}
\usage{
search.for.file(target.dir, file_name_string, full_path = FALSE)
}
\arguments{
\item{target.dir}{target directory}

\item{file_name_string}{e.g. "data_U5MR"}

\item{full_path}{full path or not}
}
\description{
Search for files containing the `file_name_string` in all sub-folders in
`target.dir`, and list files containing the `file_name_string`
}
